Refactoring: splitting component into smaller files
